feat(client): added full implementation of getCoordinates + tests + changed Element -> ScreenElement
This commit adds full implementation of getCoordinates, using a ScreenProcessor. Added full
implementation of the default ClaudeScreenProcessor, as well as a ScreenProcessorFactory that allows
for adding new processors as well as letting the developer dynamically add their own custom
processor. Added full tests for the getCoordinates method. This update also changes the Element
interface name to ScreenElement, to not clash with the base Element interface that is quite common.